1.30  武   15.50
openHB=>publishedHB
