+++ Weitianshi/project.config.json
@@ -25,7 +25,7 @@
 			"list": []
 		},
 		"miniprogram": {
-			"current": 6,
+			"current": 8,
 			"list": [
 				{
 					"id": 0,
@@ -65,26 +65,8 @@
 				},
 				{
 					"id": 6,
-					"name": "开过的红包",
+					"name": "开过红包",
 					"pathName": "redPackets/pages/openedHB/openedHB",
-					"query": ""
-				},
-				{
-					"id": -1,
-					"name": "发过的列表",
-					"pathName": "redPackets/pages/myHB/myHB",
-					"query": ""
-				},
-				{
-					"id": -1,
-					"name": "发布红宝",
-					"pathName": "redPackets/pages/publishHB/publishHB",
-					"query": ""
-				},
-				{
-					"id": -1,
-					"name": "发布红包",
-					"pathName": "redPackets/pages/publishHB/publishHB",
 					"query": ""
 				},
 				{
@@ -94,8 +76,26 @@
 					"query": ""
 				},
 				{
-					"id": -1,
+					"id": 8,
 					"name": "打开红包",
+					"pathName": "redPackets/pages/publishedHB/publishedHB",
+					"query": ""
+				},
+				{
+					"id": -1,
+					"name": "群详情",
+					"pathName": "redPackets/pages/crowdDetail/crowdDetail",
+					"query": ""
+				},
+				{
+					"id": -1,
+					"name": "开红包",
+					"pathName": "redPackets/pages/openedHB/openedHB",
+					"query": ""
+				},
+				{
+					"id": -1,
+					"name": "开红包",
 					"pathName": "redPackets/pages/openHB/openHB",
 					"query": ""
 				},
@@ -103,6 +103,12 @@
 					"id": -1,
 					"name": "群详情",
 					"pathName": "redPackets/pages/crowdDetail/crowdDetail",
+					"query": ""
+				},
+				{
+					"id": -1,
+					"name": "开红包",
+					"pathName": "redPackets/pages/openHB/openHB",
 					"query": ""
 				}
 			]
