--- Weitianshi/redPackets/pages/openHB/openHB.json
@@ -1,5 +1,0 @@
-{
-  "navigationBarTitleText": "微天使社交红包",
-  "navigationBarBackgroundColor": "#f14b50",
-  "navigationBarTextStyle": "#ffe2b1"
-}
