+++ Weitianshi/redPackets/pages/publishedHB/publishedHB.json
@@ -1,1 +1,5 @@
-{}
+{
+  "navigationBarTitleText": "微天使社交红包",
+  "navigationBarBackgroundColor": "#f14b50",
+  "navigationBarTextStyle": "#ffe2b1"
+}
