+++ Weitianshi/redPackets/pages/publishedHB/publishedHB.wxss
@@ -1,1 +1,39 @@
-/* pages/publishHB/publishHB.wxss */
+/* pages/openHB/openHB.wxss */
+
+.openZhuan {
+  width: 636rpx;
+  height: 118rpx;
+  position: fixed;
+  left: 8%;
+  bottom: 50rpx;
+}
+
+.modalBox {
+  background: rgba(0, 0, 0, 0.5);
+  width: 100%;
+  height: 100%;
+  position: absolute;
+  left: 0;
+  top: 0;
+}
+
+.tankuang {
+  height: 614rpx;
+  width: 562rpx;
+  margin: 252rpx 0 64rpx 114rpx;
+}
+
+.btnSure {
+  height: 108rpx;
+  width: 532rpx;
+  /* margin-left: 110rpx; */
+}
+
+.getMoney {
+  font-size: 80rpx;
+  color: #e80d27;
+  text-align: center;
+  position: absolute;
+  top: 470rpx;
+  left: 238rpx;
+}
