+++ Weitianshi/utils/wxss/HB.wxss
@@ -4,7 +4,7 @@
   width: 91%;
   margin: 0 32rpx;
   height: 108rpx;
-  border-bottom: 1rpx solid #fff;
+  border-bottom: 1rpx solid rgba(255, 255, 255, 0.6);
   padding: 32rpx 0;
   align-items: center;
 }
